requirements.txt
flask
groq
PyGithub
python-dotenv
fpdf2
gunicorn